Static algebraic methods passed to Algebra
